+++ src/static/publish/write-static-manifest.js
@@ -3,7 +3,7 @@ let { join } = require('path')
 let { fingerprint: fingerprinter } = require('@architect/utils')
 
 /**
- * Write (or remove) fingerprinted static asset manifest if not run as a full deploy
+ * Write, reuse, or possibly remove fingerprinted static asset manifest
  */
 module.exports = function maybeWriteStaticManifest (params, callback) {
   let { fingerprint, ignore, isFullDeploy, publicDir } = params
@@ -12,13 +12,13 @@ module.exports = function maybeWriteStaticManifest (params, callback) {
   let staticFileExists = existsSync(staticFile)
   let useExistingStaticManifest = isFullDeploy && fingerprint && staticFileExists
 
+  // Assumes static manifest was already written to disk by deploy
   if (useExistingStaticManifest) {
-    // Use the static manifest already written to disk if run as a full deploy
     let manifest = JSON.parse(readFileSync(staticFile))
     callback(null, manifest)
   }
+  // Let the fingerprinter sort out writing the manifest
   else {
-    // Let the fingerprinter sort it out
     fingerprinter({fingerprint, ignore}, callback)
   }
 }